+++ HW7/db.json
@@ -15,7 +15,7 @@
     {
       "id": 3,
       "title": "Task 3",
-      "isCompleted": false,
+      "isCompleted": true,
       "tag": "health"
     }
   ]
