+++ HW7/src/components/header/AddNewItemPopup.css
@@ -1,0 +1,71 @@
+input::placeholder {
+    color: #838383;
+}
+.divPopup {
+    display: flex;
+    align-items: center;
+    background-color: white;
+    border-radius: 24px;
+    width: 524px;
+    flex-direction: column;
+    position: absolute;
+    margin-top: 20%;
+    margin-left: 30%;
+}
+.divPopupButtons {
+    display: flex;
+    justify-content: space-around;
+    width: 524px;
+    background: #FFFFFF;
+    border-radius: 24px;
+    margin: 150px;
+    margin-bottom: 22px;
+}
+.popupCancelButton {
+    width: 227.5px;
+    height: 52px;
+    border-radius: 12px;
+    padding: 16px 32px;
+    gap: 10px;
+    background: #FFFFFF;
+    color: #0053CF;
+    font-weight: 800;
+    font-size: 16px;
+    line-height: 20px;
+    border: none;
+}
+.popupSubmitButton {
+    width: 227.5px;
+    height: 52px;
+    border-radius: 12px;
+    padding: 16px 32px;
+    gap: 10px;
+    background: #3C86F4;
+    border-radius: 12px;
+    color: #FFFFFF;
+    font-weight: 800;
+    font-size: 16px;
+    line-height: 20px;
+    border: none;
+}
+.popupSubmitButton:disabled  {
+    background: #D3D3D3;
+}
+.textInput {
+    color: #000000;
+    font-weight: 500;
+    font-size: 16px;
+    line-height: 20px;
+    box-sizing: border-box;
+    padding: 16px 21px;
+    margin-right: 40px;
+    gap: 10px;
+    width: 465px;
+    height: 52px;
+    background: #F5F5F5;
+    border: 1px solid #D2D2D2;
+    border-radius: 12px;
+    outline: none;
+    margin-right: 0px;
+    align-self:center;
+}
