+++ HW7/src/components/header/TagForm.css
@@ -1,0 +1,5 @@
+.form {
+    margin-top: 29px;
+    width: 60%;
+    margin-right: 130px;
+}
