+++ HW7/src/components/lists/App.css
@@ -1,0 +1,13 @@
+.mainDiv {
+    margin-top: 44px;
+    margin-left: 164px;
+}
+.listsDiv {
+    width: 482px;
+    margin-top: 42px;
+}
+.listHeader {
+    font-weight: 700;
+    font-size: 24px;
+    line-height: 33px;
+}
