+++ HW7/src/components/lists/DoneList.css
@@ -1,0 +1,4 @@
+.appList {
+    list-style-type: none;
+    padding-left: 12px;
+}
