+++ HW7/src/components/lists/ListElement.css
@@ -1,0 +1,61 @@
+.deleteButton {
+    border: none;
+    background: #FFFFFF;
+    float: right;
+}
+.deleteImage svg{
+    fill: #838383;
+}
+.deleteImage svg:hover {
+    fill: rgba(60, 134, 244, 0.6);
+}
+.deleteImage svg:active {
+    fill: #0053CF;
+}
+.taskElement {
+    display: flex;
+    padding: 8px 12px;
+    gap: 20px;
+}
+.spanTask {
+    font-weight: 400;
+    font-size: 18px;
+    line-height: 25px;
+    margin-left: 20px;
+    color: #1D1D1D;
+}
+.checkboxImg {
+    margin-top: 18px;
+}
+.taskDiv {
+    display: inline-block;
+    flex-grow: 1;
+}
+.tag {
+    font-family: 'Inter';
+    padding: 2px 8px;
+    border-radius: 6px;
+    font-style: normal;
+    font-weight: 500;
+    font-size: 12px;
+    line-height: 15px;
+    margin-left: 20px;  
+    display: inline-block;
+}
+.healthTag {
+    background: rgba(60, 134, 244, 0.31);
+    color: #0053CF;
+}
+.homeTag {
+    background: #E2F7E2;
+    color: #639462;
+
+}
+.workTag {
+    background: #E8D7FF;
+    color: #9747FF;
+}
+.otherTag {
+    background: #FFECC7;
+    color: #EA8C00;
+}
