+++ HW7/src/main.css
@@ -5,7 +5,7 @@ body, html {
     padding: 0;
     font-family: 'Nunito','Open Sans', sans-serif;
 }
-h1 {
+/* h1 {
     color: #1D1D1D;
     font-weight: 700;
     font-size: 32px;
